Crate re_case

Case conversions, the way Rerun likes them.

Functions

to_human_case

Converts a snake or pascal case input into “human case” output, i.e. start with upper case and continue with lower case.

to_pascal_case

Converts a snake or pascal case input into a pascal case output.

to_snake_case

Converts a snake or pascal case input into a snake case output.